Fix Pget_bkpt_ip() return value.
This function returns an arbitrary pointer, so in order to be able to
represent addresses in the upper half of the address space it must
return a uintptr_t, not a long.  Two unlikely addresses remain
unrepresentable: zero (no breakpoint) and (uintptr_t)-1 (error).

All arch implementations already conform to this contract: all we need
to do is fix a few declarations.

+++ libproc/i386/isadep.h
@@ -23,7 +23,7 @@
  * CDDL HEADER END
  */
 /*
- * Copyright 2013 Oracle, Inc.  All rights reserved.
+ * Copyright 2013, 2015 Oracle, Inc.  All rights reserved.
  * Use is subject to license terms.
  */
 
@@ -47,7 +47,7 @@ isa_dispatch_t dispatch[] = {
 
 #ifdef WANT_GET_BKPT_IP
 
-extern	long Pget_bkpt_ip_x86(struct ps_prochandle *P, int expect_esrch);
+extern	uintptr_t Pget_bkpt_ip_x86(struct ps_prochandle *P, int expect_esrch);
 
 isa_dispatch_t dispatch[] = {
     {B_TRUE, EM_X86_64, (dispatch_fun_t *) Pget_bkpt_ip_x86},
